[#7] ♻️ refactor : @Transactional 추가, 줄맞춤, cartItem 디렉토리 삭제(cart로 이동
