+++ src/main/java/com/sparta/hotitemcollector/domain/cart/CartController.java
@@ -8,47 +8,43 @@ import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.sparta.hotitemcollector.domain.cartitem.dto.CartItemResponseDto;
+import com.sparta.hotitemcollector.domain.cart.dto.CartItemResponseDto;
 import com.sparta.hotitemcollector.domain.security.UserDetailsImpl;
 import com.sparta.hotitemcollector.global.common.CommonResponse;
 
 import lombok.RequiredArgsConstructor;
 
 @RestController
+@RequestMapping("/cart")
 @RequiredArgsConstructor
 public class CartController {
 	private final CartService cartService;
 
-	@PostMapping("/carts/{productId}")
+	@PostMapping("/{productId}")
 	public ResponseEntity<CommonResponse> createCartItem(@PathVariable("productId") Long productId,
 		@AuthenticationPrincipal UserDetailsImpl userDetails) {
-
 		CartItemResponseDto responseDto = cartService.createCartItem(userDetails.getUser(), productId);
 		CommonResponse response = new CommonResponse<>("장바구니에 상품 담기 성공", 201, responseDto);
 		return new ResponseEntity<>(response, HttpStatus.CREATED);
 	}
 
-	@DeleteMapping("/carts/{productId}")
+	@DeleteMapping("/{productId}")
 	public ResponseEntity<CommonResponse> deleteCartItem(@PathVariable("productId") Long productId,
 		@AuthenticationPrincipal UserDetailsImpl userDetails) {
-
 		cartService.deleteCartItem(userDetails.getUser(), productId);
 		CommonResponse response = new CommonResponse<>("장바구니 상품 제거 성공", 204, "");
 		return new ResponseEntity<>(response, HttpStatus.OK);
 	}
 
-	@GetMapping("/carts")
+	@GetMapping
 	public ResponseEntity<CommonResponse> getCart(@RequestParam(defaultValue = "1") int page,
 		@AuthenticationPrincipal UserDetailsImpl userDetails) {
-
 		Page<CartItemResponseDto> responseDtoList = cartService.getCart(page, userDetails.getUser());
-
 		CommonResponse response = new CommonResponse<>("장바구니 조회 성공", 200, responseDtoList);
-
 		return new ResponseEntity<>(response, HttpStatus.OK);
-
 	}
 }
