+++ src/main/java/com/sparta/hotitemcollector/domain/cart/CartItem.java
@@ -1,0 +1,42 @@
+package com.sparta.hotitemcollector.domain.cart;
+
+import com.sparta.hotitemcollector.domain.product.Product;
+import com.sparta.hotitemcollector.global.Timestamped;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.Table;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@Entity
+@Table(name = "cart_item")
+@NoArgsConstructor
+public class CartItem extends Timestamped {
+	@Id
+	@GeneratedValue(strategy = GenerationType.IDENTITY)
+	@Column
+	private Long id;
+
+	@ManyToOne
+	@JoinColumn(name = "product_id", nullable = false)
+	private Product product;
+
+	@ManyToOne
+	@JoinColumn(name = "cart_id", nullable = false)
+	private Cart cart;
+
+	@Builder
+	public CartItem(Long id, Product product, Cart cart) {
+		this.id = id;
+		this.product = product;
+		this.cart = cart;
+	}
+}
